Unit tests for Post repository

+++ Domain/Repositories/IPostRepository.cs
@@ -5,5 +5,5 @@ namespace Domain.Repositories;
 public interface IPostRepository
 {
     Task<Post?> GetByIdAsync(long id);
-    Task<int> AddAsync(Post post);
+    Task<long> AddAsync(Post post);
 }
